+++ ricaun.Revit.DB.Shape.Tests/Shape_Cone_Tests.cs
@@ -4,16 +4,16 @@ using ricaun.Revit.DB.Shape.Tests.Utils;
 
 namespace ricaun.Revit.DB.Shape.Tests
 {
-    public class Shape_Pointer_Tests
+    public class Shape_Cone_Tests
     {
         [Test]
-        public void CreatePointer()
+        public void CreateCone()
         {
             var radius = 1.0;
             var center = XYZ.Zero;
-            var solid = ShapeCreator.CreatePointer(center, radius);
+            var solid = ShapeCreator.CreateCone(center, radius);
 
-            AssertUtils.Pointer(solid, radius);
+            AssertUtils.Cone(solid, radius);
         }
 
         [TestCase(1.0)]
@@ -21,12 +21,12 @@ namespace ricaun.Revit.DB.Shape.Tests
         [TestCase(3.0)]
         [TestCase(4.2)]
         [TestCase(9.8)]
-        public void CreatePointer_Radius(double radius)
+        public void CreateCone_Radius(double radius)
         {
             var center = XYZ.Zero;
-            var solid = ShapeCreator.CreatePointer(center, radius);
+            var solid = ShapeCreator.CreateCone(center, radius);
 
-            AssertUtils.Pointer(solid, radius);
+            AssertUtils.Cone(solid, radius);
         }
 
         [TestCase(1.0)]
@@ -34,13 +34,13 @@ namespace ricaun.Revit.DB.Shape.Tests
         [TestCase(3.0)]
         [TestCase(4.2)]
         [TestCase(9.8)]
-        public void CreatePointer_Height(double height)
+        public void CreateCone_Height(double height)
         {
             var radius = 1.0;
             var center = XYZ.Zero;
-            var solid = ShapeCreator.CreatePointer(center, radius, height);
+            var solid = ShapeCreator.CreateCone(center, radius, height);
 
-            AssertUtils.Pointer(solid, radius, height);
+            AssertUtils.Cone(solid, radius, height);
         }
 
         [TestCase(1.0, 1.0)]
@@ -48,12 +48,12 @@ namespace ricaun.Revit.DB.Shape.Tests
         [TestCase(3.0, 1.0)]
         [TestCase(4.2, 1.0)]
         [TestCase(9.8, 1.0)]
-        public void CreatePointer_RadiusHeight(double radius, double height)
+        public void CreateCone_RadiusHeight(double radius, double height)
         {
             var center = XYZ.Zero;
-            var solid = ShapeCreator.CreatePointer(center, radius, height);
+            var solid = ShapeCreator.CreateCone(center, radius, height);
 
-            AssertUtils.Pointer(solid, radius, height);
+            AssertUtils.Cone(solid, radius, height);
         }
     }
 }